% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cohort_covariate.R
\name{add_cohort_covariate}
\alias{add_cohort_covariate}
\title{Add a covariate variable to a USRDS cohort}
\usage{
add_cohort_covariate(
  USRDS_cohort,
  covariate_data_frame,
  covariate_date,
  covariate_value = NULL,
  covariate_variable_name
)
}
\arguments{
\item{USRDS_cohort}{is a USRDS cohort created using the \code{create_usrds_cohort()} function}

\item{covariate_data_frame}{is a data frame with a date variable (delineated via the \code{covariate_date} option).}

\item{covariate_date}{is a character string naming a date variable in
\code{covariate_data_frame} that specifies the date of onset of the covariate}

\item{covariate_value}{is an optional variable which encodes the value taken
by the time-dependent covariate on the date encoded in \code{covariate_date}.  If
\code{covariate_value} is not specified,}

\item{covariate_variable_name}{is a string with the variable name for the new covariate variable created}
}
\value{
A USRDS longitudinal cohort with additional rows created as needed
and a new time-varying covariate column named \code{covariate_variable_name}.
}
\description{
This function accepts two data frames.  One is a USRDS longitudinal cohort.  The second is a data frame with
dates representing covariates.  The resulting output is a USRDS longitudinal cohort with the covariates added to
the data frame as time-varying covariates.
}
\details{
Both \code{USRDS_cohort} and \code{condition_data_frame} must contain \code{USRDS_ID}.

Examples of characteristics
\itemize{
\item That a patient has developed a specific medical condition
\item That a patient has, in the past, experienced an event of interest
\item Patient ZIP code
}
}
\examples{
\dontrun{
}

}
