% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cohort_event.R
\name{add_cohort_event}
\alias{add_cohort_event}
\title{Add an event variable to a USRDS cohort}
\usage{
add_cohort_event(
  USRDS_cohort,
  event_data_frame,
  event_date,
  event_variable_name
)
}
\arguments{
\item{USRDS_cohort}{is a USRDS cohort created using the \code{create_usrds_cohort()} function}

\item{event_data_frame}{is a data frame with a date variable (delineated via the \code{event_date} option).}

\item{event_date}{is a character string naming a date variable in
\code{event_data_frame} that specifies when each patient experiences the event}

\item{event_variable_name}{is a string with the variable name for the new event variable created}
}
\value{
A data frame with columns \code{cohort_start_date} and \code{cohort_end_date} for each row
}
\description{
This function accepts two data frames.  One is a USRDS longitudinal cohort.  The second is a data frame with
dates representing an event.  The resulting output is a USRDS longitudinal cohort with the events added to
the data frame
}
\examples{
\dontrun{
}
}
