% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_cost_for_inflation.R
\name{adjust_costs_for_inflation}
\alias{adjust_costs_for_inflation}
\title{Adjust costs for inflation}
\usage{
adjust_costs_for_inflation(cost_data_frame, baseline_month, baseline_year)
}
\arguments{
\item{cost_data_frame}{data frame of cost data created by a \verb{get_*_*_*costs} function}

\item{baseline_month}{is the full name of the month.}

\item{baseline_year}{is a year in the 2006-2025 range.}
}
\value{
A data frame with costs adjusted for inflation, with the original variable name changed.
}
\description{
Given a data frame of claims with cost data and a month and year, adjust all costs to that month and year using CPI-M.
}
\details{
The original cost variables will be retained, and new variables created that contain "_ADJUSTED" appended to them.
}
\examples{
\dontrun{
adjust_costs_for_inflation()
}

}
