% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/establish_dx_date.R
\name{establish_dx_date}
\alias{establish_dx_date}
\title{Establish date a diagnosis is made}
\usage{
establish_dx_date(df, diagnosis_established = "diagnosis")
}
\arguments{
\item{df}{data frame of claims created by a \verb{get_*_ICD} function}
}
\value{
A data frame filtered to keep the claim which establishes that a diagnosis has been made
}
\description{
Given a data frame of claims created with a \verb{get_*_ICD} command, this function
will keep the date of either the second outpatient or
first inpatient claim with that code, using the \code{HCFASAF} variable to classify claims as outpatient
or inpatient.
}
\examples{
\dontrun{
cryptococcus_icd <- c("1175", "B45")

# Retrieve claims for all patients in selected years
result <- get_IN_ICD(icd_codes = cryptococcus_icd, years = 2013:2018)

establish_dx_date(result)
}
}
