% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_IN_ICD.R
\name{get_IN_ICD}
\alias{get_IN_ICD}
\title{Searches Institutional files for ICD codes in specific years}
\usage{
get_IN_ICD(ICDcodelist, yearlist)
}
\arguments{
\item{ICDcodelist}{ICD codes to identify}

\item{yearlist}{Years to search}
}
\value{
Data frame
}
\description{
This function searches all Institutional files from the appropriate year
}
