% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_PS_CLM_costs.R
\name{get_PS_CLM_costs}
\alias{get_PS_CLM_costs}
\title{Retrieve claims costs from Physician/Supplier claims}
\usage{
get_PS_CLM_costs(years, usrds_ids = NULL, remove_missing_values = TRUE)
}
\arguments{
\item{years}{Integer vector of calendar years to include.  Range is 2013-2021.}

\item{usrds_ids}{Optional. Vector of USRDS_IDs to restrict results to.}
}
\value{
A data frame with columns: \code{USRDS_ID}, \code{HCFASAF}, \code{CLM_FROM}, \code{CLM_THRU}, \code{CLM_PMT_AMT},
\code{CARR_CLM_PRMRY_PYR_PD_AMT},\code{NCH_CLM_PRVDR_PMT_AMT},\code{NCH_CLM_BENE_PMT_AMT},\code{NCH_CARR_SBMT_CHRG_AMT},
\code{NCH_CARR_ALOW_CHRG_AMT},\code{CARR_CLM_CASH_DDCTBL_APPLY_AMT}
}
\description{
Extracts all claims costs from physician/supplier billing data for the selected years,
optionally filtering by USRDS_IDs.
Option to \code{remove_missing_values} defaults to TRUE
}
\examples{
\dontrun{
get_PS_CLM_costs(years = 2016:2018, usrds_ids=1:1000)
get_PS_CLM_costs(years = 2014:2017, usrds_ids=2000:4100)
}
}
