% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_PS_REV_costs.R
\name{get_PS_REV_costs}
\alias{get_PS_REV_costs}
\title{Retrieve line item costs from Physician/Supplier claims}
\usage{
get_PS_REV_costs(years, usrds_ids = NULL, remove_missing_values = TRUE)
}
\arguments{
\item{years}{Integer vector of calendar years to include.}

\item{usrds_ids}{Optional. Vector of USRDS_IDs to restrict results to.}
}
\value{
A data frame with columns: \code{USRDS_ID}, \code{HCFASAF}, \code{CLM_FROM}, \code{CLM_THRU},\code{SBMTCH}, \code{ALOWCH}, \code{PMTAMT}
}
\description{
Extracts all line-item costs from physician/supplier billing data for the selected years,
optionally filtering by USRDS_IDs.
Option to \code{remove_missing_values} defaults to TRUE
}
\examples{
\dontrun{
get_PS_REV_costs(years = 2006:2008, usrds_ids=1:1000)
get_PS_REV_costs(years = 2006:2008, usrds_ids=2000:4100)
}
}
