% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_PS_specialty.R
\name{get_PS_specialty}
\alias{get_PS_specialty}
\title{Extract PS specialty claims}
\usage{
get_PS_specialty(
  specialtycodelist,
  yearlist,
  DIAG = FALSE,
  HCPCS = FALSE,
  PLCSRV = FALSE
)
}
\arguments{
\item{specialtycodelist}{A vector of specialty codes to keep.}

\item{yearlist}{A vector of years to include.}

\item{DIAG}{Logical, include DIAG column.}

\item{HCPCS}{Logical, include HCPCS column.}

\item{PLCSRV}{Logical, include PLCSRV column.}
}
\value{
A tibble with filtered and selected PS HCPCS data.
}
\description{
Returns claims filtered by specialty code from the PS HCPCS dataset.
}
