% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prorate_costs_by_day.R
\name{prorate_costs_by_day}
\alias{prorate_costs_by_day}
\title{Prorate costs by day}
\usage{
prorate_costs_by_day(cost_data_frame)
}
\arguments{
\item{cost_data_frame}{data frame of cost data created by a \verb{get_*_*_*costs} function}
}
\value{
A data frame with costs prorated, with the original variable name changed.
}
\description{
Given a data frame of claims with cost data from claims which potentially span multiple days,
this function will prorate the costs evenly over the duration of the claim.
}
\details{
The original cost variables will be removed, and replaced by variables with the same name but with
"_PRORATED" added to the variable na,e.
}
\examples{
\dontrun{
prorate_costs_by_day()
}

}
