% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_medicare_primary.R
\name{verify_medicare_primary}
\alias{verify_medicare_primary}
\title{Verify that a patient has Medicare primary coverage}
\usage{
verify_medicare_primary(
  df,
  index_date,
  lookback_days = 365,
  medicare_coverage_df = NULL
)
}
\arguments{
\item{df}{is a data frame, presumably one with a single row per patient.  Data frame \emph{must} contain
a column named \code{USRDS_ID}.}

\item{index_date}{either a fixed date or the name of a date variable in quotes}

\item{lookback_days}{is the number of days of Medicare as primary necessary.  Default value is 365.}

\item{medicare_coverage_df}{is an optional parameter containing a df from the \code{payhist} file.}
}
\value{
The original \code{df} data fame with the variable \code{medicare_primary_TF} added
}
\description{
This function accepts a data frame \code{df}, an index date option \code{index_date} which is either a fixed date or
the name of a date variable in quotes, and a variable \code{lookback_days} that encodes how many days of Medicare
coverage are necessary.
}
\details{
An optional variable \code{medicare_coverage_df} allows for a df containing information from the \code{payhist} file to be passed to the function.  If left
NULL, then the function will load the data via the \code{load_usrds_file} function.

It returns the same dataframe with a new variable \code{medicare_primary_TF} that takes the values \code{TRUE}
or \code{FALSE}.
}
\examples{
\dontrun{

test_df<-data.frame(
USRDS_ID=c(30, 111112),
initial_dates=as.Date(c("1994-01-01", "2019-01-01"))
)

verify_medicare_primary(df=test_df, index_date = "initial_dates", lookback_days = 365)

}
}
