% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalize_usrds_cohort.R
\name{finalize_usrds_cohort}
\alias{finalize_usrds_cohort}
\title{Finalize a USRDS cohort}
\usage{
finalize_usrds_cohort(USRDS_cohort)
}
\arguments{
\item{USRDS_cohort}{is a USRDS cohort created using the \code{create_usrds_cohort()} function
#'}
}
\value{
A data frame that is a USRDS cohort object
}
\description{
This function accepts a data frame that is a USRDS longitudinal cohort, and returns the same object with the
\code{tstart} and \code{tstop} variables shifted so that each patient's first observation has \code{tstart}=0.
}
\examples{
\dontrun{
}
}
