% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_PS_HCPCS.R
\name{get_PS_HCPCS}
\alias{get_PS_HCPCS}
\title{Searches Physician Supplier files for HCPCS codes in specific years}
\usage{
get_PS_HCPCS(HCPCScodelist, yearlist)
}
\arguments{
\item{HCPCScodelist}{ICD codes to identify}

\item{yearlist}{Years to search}
}
\value{
Data frame
}
\description{
This function searches all physician supplier files from the appropriate year
}
