% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_PS_ICD.R
\name{get_PS_ICD}
\alias{get_PS_ICD}
\title{Searches Physician Supplier files for ICD codes in specific years}
\usage{
get_PS_ICD(ICDcodelist, yearlist)
}
\arguments{
\item{ICDcodelist}{ICD codes to identify}

\item{yearlist}{Years to search}
}
\value{
Data frame
}
\description{
This function searches all physician supplier files from the appropriate year
}
