% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_USRDS_wd.R
\name{set_USRDS_wd}
\alias{set_USRDS_wd}
\title{Set USRDS working directory}
\usage{
set_USRDS_wd(x)
}
\arguments{
\item{x}{The path to the directory where the USRDS files are stored}
}
\description{
Set USRDS working directory
}
