% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usRds-package.R
\docType{package}
\name{usRds-package}
\alias{usRds}
\alias{usRds-package}
\title{usRds: A package to automate processes involving the USRDS}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/VagishHemmige/usRds}
  \item \url{https://vagishhemmige.github.io/usRds/}
  \item Report bugs at \url{https://github.com/VagishHemmige/usRds/issues}
}

}
\author{
\strong{Maintainer}: Vagish Hemmige \email{vahemmig@montefiore.org}

}
\keyword{internal}
